Fix single tap bug
There was a bug where if you tapped on a new tile, didn't drag it, then released it, that tile would swap with the last one you were dragging. This was because the game still thought your cursor position was at the old tile, because it only updates the cursor position on movement.